feat: add comprehensive interview knowledge base with 12 categories, code examples & .gitignore
- Added interview-data.js (1725 lines) with 100+ Q&As covering Java/Spring Boot,
  Microservices, Design Patterns, Resilience, Database, Security, DevOps/Docker/K8s,
  Frontend React/Angular, Observability, Testing, System Design, Behavioral
- Updated app.js with tabbed interview view (Comprehensive + Quick Reference)
- Added KB CSS styles for categories, topics, code blocks
- Added script tag for interview-data.js in index.html
- Created project-wide .gitignore (Java, Node, IDE, Docker, Terraform, etc.)
- Includes all recent microservice enhancements and architecture updates

+++ employee-microservice/src/main/java/com/example/employee/anticorruption/LegacyPayrollSystemDTO.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 /**
  * Anti-Corruption Layer (ACL) Pattern
  * 
@@ -28,7 +30,7 @@ public class LegacyPayrollSystemDTO {
     private String emailAddr;          // Instead of email
     private String dept;               // Instead of department
     private String pos;                // Instead of position
-    private Double monthlySal;         // Instead of salary
+    private BigDecimal monthlySal;     // Instead of salary
     private String stat;               // Instead of status
     private String hiredDate;          // Instead of hireDate (String format)
