+++ employee-microservice/src/main/java/com/example/employee/client/PayrollResponse.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -12,7 +13,7 @@ import java.time.LocalDate;
 public class PayrollResponse {
     private Long id;
     private Long employeeId;
-    private Double salary;
+    private BigDecimal salary;
     private String status;
     private LocalDate effectiveDate;
 }
